LoginService Update

+++ src/main/java/net/ddns/sbapiserver/service/login/LoginService.java
@@ -2,6 +2,9 @@ package net.ddns.sbapiserver.service.login;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import net.ddns.sbapiserver.common.code.ErrorCode;
+import net.ddns.sbapiserver.exception.error.custom.BusinessException;
+import net.ddns.sbapiserver.util.JwtUtil;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 
@@ -13,12 +16,8 @@ import java.util.concurrent.TimeUnit;
 public class LoginService {
 
     private final RedisTemplate<String, String> redisTemplate;
+    private final JwtUtil jwtUtil;
     private static final int ACCESS_TOKEN_EXPIRATION_TIME = 30 * 60;
-    private static final int EXPIRED_TOKEN_TIME = 0;
-
-    public void handleConcurrentLogin(String userId, String accessToken){
-        redisTemplate.opsForValue().get(userId)
-    }
 
     // Redis에 AccessToken 저장
     public void storeAccessToken(String userId, String token){
@@ -31,13 +30,22 @@ public class LoginService {
         redisTemplate.delete(userId);
     }
 
+    public boolean isUserLoginValid(String userid, String accessToken){
+        String findToken = redisTemplate.opsForValue().get(userid);
+        log.error("찾은 Token : {} 입력한 Token : {}", findToken, accessToken);
+        return findToken.equals(accessToken);
+    }
+
     // Login 중 -> true, or false
     public boolean isUserLoggedIn(String userId){
         return redisTemplate.opsForValue().get(userId) != null;
     }
 
-    public void logout(String accessToken, String userId){
+    public void logout(String userId){
+        if(!isUserLoggedIn(userId)){
+            throw new BusinessException(ErrorCode.USER_NOT_LOGGED_ERROR, ErrorCode.USER_NOT_LOGGED_ERROR.getReason());
+        }
         deleteAccessToken(userId);
-        log.info("User : [{}] Logout", userId);
+        log.info("User : [{}] Logout", userId.toString());
     }
 }
